+++ module/SD/Admin/src/Admin/Entity/User.php
@@ -81,7 +81,7 @@ final class User
     /**
      * @var string
      *
-     * @ORM\Column(name="image", type="string", length=100, nullable=true)
+     * @ORM\Column(name="image", type="string", length=255, nullable=true)
      */
     private $image;