+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # CHANGELOG #
 DD.MM.YYYY
-## 0.0.23 (17.11.2015) - Stable Alpha ##
+## 0.0.23 (18.11.2015) - Stable Alpha ##
 - Fixed session collisions - The ap was creating two sessions __zpc and zpc
 - Fixed a bug, where Firefox couldn't load assests files due to misconfigured url slashes
 - First drafts of ProfileController
 - Bumped cookie life time to 10 days
 - Minor bug fixes
 - Updated composer.phar
+- Added SettingsController - this one sits in Application module and is used by every user. I could name it UserSettingsController, but I don't want to do that for now.
 
 DD.MM.YYYY
 ## 0.0.22 (9.11.2015) - Stable Alpha ##